progress! working in dev
