--- experiments/rsc/src/bootstrap/bootstrapClient.ts
@@ -1,3 +1,0 @@
-import { ssrWebpackRequire } from "../register/rsc";
-
-globalThis.__webpack_require__ = ssrWebpackRequire;
