--- experiments/rsc/src/bootstrap/bootstrapWorker.ts
@@ -1,3 +1,0 @@
-import { ssrWebpackRequire } from "../register/rsc";
-
-globalThis.__webpack_require__ = ssrWebpackRequire;
