+++ experiments/rsc/src/register/client.ts
@@ -1,6 +1,6 @@
 import { createServerReference as baseCreateServerReference } from "react-server-dom-webpack/client.browser";
 
-export function createServerReference(id: string, name: string) {
+export const createServerReference = (id: string, name: string) => {
   id = id + "#" + name;
   return baseCreateServerReference(id, globalThis.__rsc_callServer);
-}
+};
