+++ experiments/rsc/src/register/worker.ts
@@ -3,15 +3,7 @@ import {
   registerClientReference as baseRegisterClientReference,
   decodeReply,
 } from "react-server-dom-webpack/server.edge";
-import memoize from "lodash/memoize";
-
-const modules = (
-  import.meta as object as {
-    glob: (
-      pattern: string,
-    ) => Record<string, () => Promise<Record<string, unknown>>>;
-  }
-).glob("/src/app/**/*.{ts,tsx}");
+import { getModuleExport } from "../imports/worker";
 
 export function registerServerReference(
   action: Function,
@@ -35,24 +27,6 @@ export function registerClientReference(id: string, exportName: string) {
     },
   );
 }
-
-export const loadModule = memoize(async (moduleName: string) => {
-  return await modules[moduleName]();
-});
-
-export const getModuleExport = async (id: string) => {
-  const [file, name] = id.split("#");
-  const module = await loadModule(file);
-  return module[name];
-};
-
-// context(justinvdm, 2 Dec 2024): re memoize(): React relies on the same promise instance being returned for the same id
-export const ssrWebpackRequire = memoize(async (id: string) => {
-  const [file, name] = id.split("#");
-  const module = await loadModule(file);
-  return { [id]: module[`${name}SSR`] };
-});
-
 export async function rscActionHandler(req: Request): Promise<unknown> {
   const url = new URL(req.url);
   const contentType = req.headers.get("content-type");
