+++ experiments/rsc/types/global.d.ts
@@ -6,6 +6,13 @@ import { CallServerCallback } from "react-server-dom-webpack/client.browser";
 declare global {
   var __webpack_require__: (id: string) => unknown;
   var __rsc_callServer: CallServerCallback;
+  interface ImportMeta {
+    readonly env: ImportMetaEnv;
+  }
+
+  interface ImportMetaEnv {
+    readonly DEV: boolean;
+  }
 }
 
 export {};
